Moving all model modules inside the recipe for not messing with the core code as specified. However, in a next PR we can move the two-step tdcn inside the main block codebase as it is almost ready
